pyproject.toml
[project]
name = "llmgusser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nonomura", email = "kanakana.nono1224@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "llama-cpp-python>=0.3.9",
]

[project.scripts]
llmgusser = "llmgusser.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
